requirements.txt
openai
pandas
openpyxl
python-dotenv